requirements.txt
# 舆情分析系统依赖包
# Spark相关
pyspark>=3.5.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 中文分词（可选，如果需要更高级的分词）
# jieba>=0.42.1

# 可视化（可选）
matplotlib>=3.7.0
seaborn>=0.12.0

# 数据库连接（如果使用MySQL）
# pymysql>=1.1.0

# 其他工具
pyyaml>=6.0

